requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
Brotli==1.0.9
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
colorama==0.4.4
colorlover==0.3.0
cryptography==3.1
cufflinks==0.17.3
cycler==0.10.0
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dnspython==1.16.0
dom==0.9
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
fuzzywuzzy==0.18.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.29
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kaleido==0.0.3.post1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mstranslate==1.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.5
numpy==1.19.1
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.12.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-Levenshtein-wheels==0.13.1
python-telegram-bot==12.8
pytz==2020.4
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
requests==2.24.0
retrying==1.3.3
rfc3986==1.4.0
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.5
urllib3==1.25.10
virtualenv==20.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yandex-geocoder==2.0.0
